+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 - Integrate Camel-Spring-Boot version 3.10.0.
 - Provide REST endpoints for adding and removing Camel routes and Spring beans at runtime.
 - Add Camel error handler for propagating errors in routes to the Configuration Manager.
-- Send ArtifactRequest and ArtifactResponse message to Clearing House.
+- Send `ArtifactRequest` and `ArtifactResponse` messages to the Clearing House.
 - Allow artifacts pointing to backend systems to be created with both BasicAuth and API key
   authentication.